Renamed some components for consistency
