+++ examples/src/app/AppRouterAnt.tsx
@@ -3,7 +3,7 @@ import * as React from 'react';
 import { NavigationAnt } from '../common/NavigationAnt';
 
 import { LocationManager } from '@moxb/moxb';
-import { LocationDependentContentImpl } from '@moxb/antd';
+import { LocationDependentArea } from '@moxb/antd';
 
 import { defaultContent, mainMenu } from '../MenuStructure';
 
@@ -15,9 +15,9 @@ export class AppRouterAnt extends React.Component<{ location?: LocationManager }
                     <Row>
                         <NavigationAnt />
                     </Row>
-                    <LocationDependentContentImpl subStates={mainMenu} fallback={defaultContent} part="main" />
+                    <LocationDependentArea subStates={mainMenu} fallback={defaultContent} part="main" />
                     <hr />
-                    <LocationDependentContentImpl subStates={mainMenu} fallback={defaultContent} part="bottom" />
+                    <LocationDependentArea subStates={mainMenu} fallback={defaultContent} part="bottom" />
                 </Layout.Content>
             </Layout>
         );
