+++ packages/antd/src/not-antd/index.ts
@@ -1,5 +1,4 @@
 export { Link } from './Link';
 export { ArgChangingLink } from './ArgChangingLink';
 export { MultiArgChangingLink } from './MultiArgChangingLink';
-export { LocationDependentContentProps } from './LocationDependentContent';
-export { LocationDependentContentImpl } from './LocationDependentContentImpl';
+export { LocationDependentArea, LocationDependentAreaProps } from './LocationDependentArea';
